final_project/requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
Appium-Python-Client==2.2.0
astroid==2.11.4
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
dill==0.3.4
enum34==1.1.10
execnet==1.9.0
Faker==13.7.0
flake8==4.0.1
Flask==2.1.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
lxml==4.8.0
MarkupSafe==2.1.1
mccabe==0.6.1
namedlist==1.8
outcome==1.1.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.13.8
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.1
pytest-forked==1.4.0
pytest-repeat==0.9.1
pytest-xdist==2.5.0
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.8
webdriver-manager==3.5.3
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
